# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智能墨水屏平台 API
  version: 1.0.0
  description: |
    用户端/管理端/设备端统一接口文档。
    设备（真机与模拟）推荐统一启动链路：login -> auto-register -> bind/status -> login -> stream。
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/hardware/tf/download/{fileId}:
    get:
      summary: 设备下载TF文件
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: file stream
...
  /api/tf/{fileId}/download:
    get:
      summary: 下载TF卡文件
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: fileId
          required: true
          schema: { type: string }
      responses:
        '200':
          description: file stream
  /api/remote/show-text:
    post:
      summary: 远程显示文本
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                deviceIds: { type: array, items: { type: string } }
                clusterIds: { type: array, items: { type: string } }
                randomCount: { type: integer, minimum: 1 }
                text: { type: string }
                durationSec: { type: number }
                saveToTf: { type: boolean }
      responses:
        '200':
          description: ok
  /api/remote/show-image:
    post:
      summary: 远程显示图片
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                deviceIds: { type: array, items: { type: string } }
                clusterIds: { type: array, items: { type: string } }
                randomCount: { type: integer, minimum: 1 }
                saveToTf: { type: boolean }
                file: { type: string, format: binary }
      responses:
        '200':
          description: ok
  /api/remote/switch-view:
    post:
      summary: 远程切换界面
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                deviceIds: { type: array, items: { type: string } }
                clusterIds: { type: array, items: { type: string } }
                randomCount: { type: integer, minimum: 1 }
                view: { type: string }
                saveToTf: { type: boolean }
      responses:
        '200':
          description: ok
  /api/remote/cast-frame:
    post:
      summary: 投屏帧上传
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                imageData: { type: string }
                width: { type: number }
                height: { type: number }
      responses:
        '200':
          description: ok
  /api/remote/cast-stop:
    post:
      summary: 结束投屏
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                deviceId: { type: string }
                deviceIds: { type: array, items: { type: string } }
                clusterIds: { type: array, items: { type: string } }
                randomCount: { type: integer, minimum: 1 }
                reason: { type: string }
      responses:
        '200':
          description: ok
...